pyproject.toml
[project]
name = "otools"
version = "1.0.0"
description = "Local LLM tools for Osaurus"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.33.0",
    "rich>=14.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
all = [
    "ddgs>=9.0.0",
    "beautifulsoup4>=4.12.0",
    "cryptography>=41.0.0",
    "playwright>=1.40.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "mlx-lm",
    "mlx-vlm",
]

[dependency-groups]
test = ["pytest"]